+++ Makefile
@@ -1,11 +1,16 @@
-.PHONY: run-demo unify
+.PHONY: panos unify plots
 
-run-demo:
-	@echo "Running demo script..."
-	@poetry run python -m scripts.run_demo
-	@echo "Demo completed. Check data/demo directory for output files."
+panos:
+	@echo "Running panos script..."
+	@poetry run python -m scripts.panos
+	@echo "Panos completed. Check data/demo directory for output files."
 
 unify:
 	@echo "Running point unification on panorama data..."
-	@poetry run python -m scripts.test_point_unification
+	@poetry run python -m scripts.unify
 	@echo "Point unification completed. Check data/point_unification_results directory for output files."
+
+plots:
+	@echo "Running plot generation..."
+	@poetry run python -m scripts.plots
+	@echo "Plot generation completed. Check data/plots directory for output files."
