+++ src/__init__.py
@@ -1,1 +1,1 @@
-from . import utils, sv, osm_streets
+from . import utils
