--- src/sv.py
@@ -1,25 +1,0 @@
-import pandas as pd
-from streetview import search_panoramas
-from tqdm import tqdm
-
-
-def get_panos(points_gdf, verbose=False):
-    panos = []
-    for _, row in tqdm(
-        points_gdf.iterrows(), total=len(points_gdf), disable=not verbose
-    ):
-        lat, lon = row["geometry"].y, row["geometry"].x
-        panos_there = search_panoramas(lat=lat, lon=lon)
-
-        for pano in panos_there:
-            id = pano.pano_id
-            lat, lon = pano.lat, pano.lon
-            date = pano.date
-            panos.append((id, lat, lon, date))
-
-    panos = (
-        pd.DataFrame(panos, columns=["pano_id", "lat", "lon", "date"])
-        .drop_duplicates()
-        .reset_index(drop=True)
-    )
-    return panos
